Refactor set to use LinkedList methods
This ups the time complexity (one loop thorugh the linked list to
find the index and then another to change the value, in the worst
case), but the point of a hashmap is to minimize the collisions
for a given hash, so this should have a negligible performance
impact.
Particularly considering this is an exercise and never going to be
used for a real application. You might say the performance impact
is so small it's already vanished.