Fix bugs with rebalancing, length, and correctly adding new nodes
Adding new nodes: I was incorrectly adding a node to the new map
when the map expected to take a key and value and create the node
itself. Javascript's lack of typing enables another bug :|

Length: I was increasing the length of the map even when modifying
the value of an existing key. When fixing this, I refactored a
little in a way I thought made the code more readable.

Rebalancing: I was attempting to destructure a two element array
into two arguments input to an arrow function. Either this can't
be done in JS, or I got the syntax wrong.